converter/requirements.txt
numpy==1.16.4
pandas==0.24.2
Pillow==6.0.0